finetune
git-svn-id: https://svn.potix.com/svn/zk1/trunk@15319 dd50bd9b-9913-0410-b9ba-e07a3075be86

+++ zk/src/org/zkoss/zk/ui/impl/EventProcessingThreadImpl.java
@@ -324,7 +324,7 @@ implements EventProcessingThread {
 		final Configuration config = desktop.getWebApp().getConfiguration();
 		_evtThdInits = config.newEventThreadInits(comp, event);
 		try {
-			int evtTimeWarn = config.getEventTimeWarning();
+			long evtTimeWarn = config.getEventTimeWarning();
 			long begt = 0;
 			if (evtTimeWarn > 0) {
 				begt = System.currentTimeMillis();
